Unnecessary classes were deleted
