+++ DEV-10/ArrayConsistingOfRepeatingElements/ArrayChecker.cs
@@ -1,29 +1,20 @@
 ﻿using System.Collections.Generic;
 
 namespace ArrayConsistingOfRepeatingElements
-{ 
+{
+  /// <summary>
+  /// This class checks is the result array empty or not
+  /// </summary>
+  class ArrayChecker
+  {
     /// <summary>
-    /// This class checks is the result array empty or not
+    /// Method checks the array on emptiness
     /// </summary>
-    class ArrayChecker
+    /// <param name="resultArray">the result array</param>
+    /// <returns>bool checker - the result of checking: true, if it's empty, else false. </returns>
+    public bool CheckArrayEmptiness(List<double> resultArray)
     {
-        /// <summary>
-        /// Method checks the array on emptiness
-        /// </summary>
-        /// <param name="resultArray">the result array</param>
-        /// <returns>bool checker - the result of checking: true, if it's empty, else false. </returns>
-        public bool CheckArrayEmptiness(List<double> resultArray)
-        {
-            bool checker;
-            if(resultArray.Count == 0)
-            {
-                checker = true;
-            }
-            else
-            {
-                checker = false;
-            }
-            return checker;
-        }
+      return (resultArray.Count == 0);
     }
+  }
 }
