+++ DEV-10/ArrayConsistingOfRepeatingElements/ArrayConsistingOfRepeatingElements.csproj
@@ -47,8 +47,7 @@
     <Compile Include="EntryPoint.cs" />
     <Compile Include="Outputer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RepeatingNumbersArrayCreator.cs" />
-    <Compile Include="SuperArrayCreator.cs" />
+    <Compile Include="RepeatedNumbersArrayCreator.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
